Refactor location of outputwriters. Now they are in the package io.writer
