+++ src/main/java/de/tum/bgu/msm/longDistance/AirportAnalysisModel.java
@@ -3,15 +3,9 @@ package de.tum.bgu.msm.longDistance;
 import de.tum.bgu.msm.Util;
 import de.tum.bgu.msm.longDistance.airportAnalysis.AirportAnalysis;
 import de.tum.bgu.msm.longDistance.data.DataSet;
-import de.tum.bgu.msm.longDistance.destinationChoice.DestinationChoice;
-import de.tum.bgu.msm.longDistance.destinationChoice.ZoneDisaggregator;
-import de.tum.bgu.msm.longDistance.io.OutputWriter;
+import de.tum.bgu.msm.longDistance.io.writer.OutputWriter;
 import de.tum.bgu.msm.longDistance.io.reader.SkimsReader;
-import de.tum.bgu.msm.longDistance.io.reader.SyntheticPopulationReader;
 import de.tum.bgu.msm.longDistance.io.reader.ZoneReader;
-import de.tum.bgu.msm.longDistance.modeChoice.ModeChoice;
-import de.tum.bgu.msm.longDistance.timeOfDay.TimeOfDayChoice;
-import de.tum.bgu.msm.longDistance.tripGeneration.TripGeneration;
 import org.apache.log4j.Logger;
 import org.json.simple.JSONObject;
