+++ src/main/java/de/tum/bgu/msm/longDistance/LDModelGermany.java
@@ -4,7 +4,7 @@ import de.tum.bgu.msm.Util;
 import de.tum.bgu.msm.longDistance.data.DataSet;
 import de.tum.bgu.msm.longDistance.destinationChoice.DestinationChoice;
 import de.tum.bgu.msm.longDistance.emissions.Emissions;
-import de.tum.bgu.msm.longDistance.io.OutputWriter;
+import de.tum.bgu.msm.longDistance.io.writer.OutputWriter;
 import de.tum.bgu.msm.longDistance.io.reader.EconomicStatusReader;
 import de.tum.bgu.msm.longDistance.io.reader.SkimsReader;
 import de.tum.bgu.msm.longDistance.io.reader.SyntheticPopulationReader;
