+++ src/main/java/de/tum/bgu/msm/longDistance/io/writer/OutputWriter.java
@@ -1,4 +1,4 @@
-package de.tum.bgu.msm.longDistance.io;
+package de.tum.bgu.msm.longDistance.io.writer;
 
 import de.tum.bgu.msm.longDistance.ModelComponent;
