+++ src/main/java/de/tum/bgu/msm/longDistance/io/writer/OutputWriterOntario.java
@@ -1,10 +1,11 @@
-package de.tum.bgu.msm.longDistance.io;
+package de.tum.bgu.msm.longDistance.io.writer;
 
 import de.tum.bgu.msm.JsonUtilMto;
 import de.tum.bgu.msm.Util;
 import de.tum.bgu.msm.longDistance.data.DataSet;
 import de.tum.bgu.msm.longDistance.data.trips.LongDistanceTrip;
 import de.tum.bgu.msm.longDistance.data.trips.LongDistanceTripOntario;
+import de.tum.bgu.msm.longDistance.io.writer.OutputWriter;
 import org.json.simple.JSONObject;
 
 import java.io.PrintWriter;
